Update: Migrate build system to CMake. Fix blended textures.

+++ TOMB4/game/box.h
@@ -1,42 +1,42 @@
 #pragma once
 #include "../global/types.h"
 
-void CreatureDie(short item_number, bool explode);
-void InitialiseCreature(short item_number);
-long CreatureActive(short item_number);
+void CreatureDie(int16_t item_number, bool explode);
+void InitialiseCreature(int16_t item_number);
+int32_t CreatureActive(int16_t item_number);
 void CreatureAIInfo(ITEM_INFO* item, AI_INFO* info);
-long SearchLOT(LOT_INFO* LOT, long expansion);
-long UpdateLOT(LOT_INFO* LOT, long expansion);
-void TargetBox(LOT_INFO* LOT, short box_number);
-long EscapeBox(ITEM_INFO* item, ITEM_INFO* enemy, short box_number);
-long ValidBox(ITEM_INFO* item, short zone_number, short box_number);
-long StalkBox(ITEM_INFO* item, ITEM_INFO* enemy, short box_number);
+int32_t SearchLOT(LOT_INFO* LOT, int32_t expansion);
+int32_t UpdateLOT(LOT_INFO* LOT, int32_t expansion);
+void TargetBox(LOT_INFO* LOT, int16_t box_number);
+int32_t EscapeBox(ITEM_INFO* item, ITEM_INFO* enemy, int16_t box_number);
+int32_t ValidBox(ITEM_INFO* item, int16_t zone_number, int16_t box_number);
+int32_t StalkBox(ITEM_INFO* item, ITEM_INFO* enemy, int16_t box_number);
 target_type CalculateTarget(PHD_VECTOR* target, ITEM_INFO* item, LOT_INFO* LOT);
 void CreatureMood(ITEM_INFO* item, AI_INFO* info, bool violent);
 void GetCreatureMood(ITEM_INFO* item, AI_INFO* info, bool violent);
-long CreatureCreature(short item_number);
-long BadFloor(long x, long y, long z, long box_height, long next_height, short room_number, LOT_INFO* LOT);
-long CreatureAnimation(short item_number, short angle, short tilt);
-short CreatureTurn(ITEM_INFO* item, short maximum_turn);
-void CreatureTilt(ITEM_INFO* item, short angle);
-void CreatureJoint(ITEM_INFO* item, short joint, short required);
-void CreatureFloat(short item_number);
-void CreatureUnderwater(ITEM_INFO* item, long depth);
-short CreatureEffect(ITEM_INFO* item, BITE_INFO* bite, short(*generate)(long x, long y, long z, short speed, short yrot, short room_number));
-short CreatureEffectT(ITEM_INFO* item, BITE_INFO* bite, short damage, short angle,
-	short(*generate)(long x, long y, long z, short damage, short angle, short room_number));
-long CreatureVault(short item_number, short angle, long vault, long shift);
-void CreatureKill(ITEM_INFO* item, short kill_anim, short kill_state, short lara_anim);
-void AlertAllGuards(short item_number);
+int32_t CreatureCreature(int16_t item_number);
+int32_t BadFloor(int32_t x, int32_t y, int32_t z, int32_t box_height, int32_t next_height, int16_t room_number, LOT_INFO* LOT);
+int32_t CreatureAnimation(int16_t item_number, int16_t angle, int16_t tilt);
+int16_t CreatureTurn(ITEM_INFO* item, int16_t maximum_turn);
+void CreatureTilt(ITEM_INFO* item, int16_t angle);
+void CreatureJoint(ITEM_INFO* item, int16_t joint, int16_t required);
+void CreatureFloat(int16_t item_number);
+void CreatureUnderwater(ITEM_INFO* item, int32_t depth);
+int16_t CreatureEffect(ITEM_INFO* item, BITE_INFO* bite, int16_t(*generate)(int32_t x, int32_t y, int32_t z, int16_t speed, int16_t yrot, int16_t room_number));
+int16_t CreatureEffectT(ITEM_INFO* item, BITE_INFO* bite, int16_t damage, int16_t angle,
+                        int16_t(*generate)(int32_t x, int32_t y, int32_t z, int16_t damage, int16_t angle, int16_t room_number));
+int32_t CreatureVault(int16_t item_number, int16_t angle, int32_t vault, int32_t shift);
+void CreatureKill(ITEM_INFO* item, int16_t kill_anim, int16_t kill_state, int16_t lara_anim);
+void AlertAllGuards(int16_t item_number);
 void AlertNearbyGuards(ITEM_INFO* item);
-short AIGuard(CREATURE_INFO* creature);
-void FindAITargetObject(CREATURE_INFO* creature, short obj_num);
+int16_t AIGuard(CREATURE_INFO* creature);
+void FindAITargetObject(CREATURE_INFO* creature, int16_t obj_num);
 void GetAITarget(CREATURE_INFO* creature);
-short SameZone(CREATURE_INFO* creature, ITEM_INFO* target_item);
-void CreatureYRot(PHD_3DPOS* srcpos, short angle, short angadd);
-long MoveCreature3DPos(PHD_3DPOS* srcpos, PHD_3DPOS* destpos, long velocity, short angdif, long angadd);
+int16_t SameZone(CREATURE_INFO* creature, ITEM_INFO* target_item);
+void CreatureYRot(PHD_3DPOS* srcpos, int16_t angle, int16_t angadd);
+int32_t MoveCreature3DPos(PHD_3DPOS* srcpos, PHD_3DPOS* destpos, int32_t velocity, int16_t angdif, int32_t angadd);
 
 extern BOX_INFO* boxes;
-extern ushort* overlap;
-extern short* ground_zone[5][2];
-extern long num_boxes;
+extern uint16_t* overlap;
+extern int16_t* ground_zone[5][2];
+extern int32_t num_boxes;
